+++ R/utils.R
@@ -3,3 +3,10 @@ reclass <- function(x) {
     class(x) <- c("quantities", "units", "errors")
   x
 }
+
+dfapply <- function(X, FUN, ...) {
+  attrs <- attributes(X)
+  X <- lapply(X, FUN, ...)
+  attributes(X) <- attrs
+  X
+}